\name{inducedGraph}
\alias{inducedGraph}
\title{inducedGraph}
\usage{
  inducedGraph(dag, startNodes)
}
\arguments{
  \item{dag}{graphNEL, a graphnel object of DAG network}
  \item{startNodes}{character vector, nodes are used to backtrack}
}
\value{
  graphNEL list, DAG network as graphnel class
}
\description{
  return subGraph form DAG network by startNodes
}
\details{
  return subGraph form DAG network by startNodes
}
\author{
  tmcjp
}

