\name{pctermlist}
\alias{pctermlist}
\title{pctermlist}
\usage{
  pctermlist(ontology, ontoterm, symbol_2_ID)
}
\arguments{
  \item{ontology}{ontology object}
  \item{ontoterm}{ontology all term ID}
  \item{symbol_2_ID}{character, term symbol character that name attr is term ID}
}
\value{
  list, Including parent and child node correspondence
}
\description{
  creat term list that Including parent and child node correspondence
}
\details{
  creat term list that Including parent and child node correspondence
}
\author{
  tmcjp
}
