\name{reverseArch}
\alias{reverseArch}
\title{reverseArch}
\usage{
  reverseArch(dirGraph, useAlgo = 'sparse', useWeights = TRUE)
}
\arguments{
  \item{dirGraph}{graphNEL, a graphnel object of directed network}
  \item{useAlgo}{character, which method to be used}
  \item{useWeights}{logical, Whether to use weight}
}
\value{
  graphNEL, a DAG network graphNEL that includes matrix
}
\description{
  reverse Arch the DAG graphNEL
}
\details{
  reverse Arch the DAG graphNEL
}
\author{
  tmcjp
}
