\name{symbol_2_ID}
\alias{symbol_2_ID}
\title{symbol_2_ID}
\usage{
  symbol_2_ID(ontology)
}
\arguments{
  \item{ontology}{ontology object}
}
\value{
  character, term symbol character that name attr is term ID
}
\description{
  creat term symbol character that name attr is term ID
}
\details{
  creat term symbol character that name attr is term ID
}
\author{
  tmcjp
}
