\name{visDAG}
\alias{visDAG}
\title{visDAG}
\usage{
  visDAG(DAGmatrix, symbol_2_ID, knownNodes, shapes = 'dot')
}
\arguments{
  \item{DAGmatrix}{graphNEL, a DAG network graphNEL that includes matrix}
  \item{symbol_2_ID}{character, term symbol to term ID}
  \item{knownNodes}{character vector, nodes are used for backtracking}
  \item{shapes}{character, DAG network shape}
}
\value{
  html, the html of DAG network graph
}
\description{
  creat html of DAG graph by visNetwork
}
\details{
  creat html of DAG graph by visNetwork
}
\author{
  tmcjp
}
